# Agent-Study/src/main/resources/swagger/openAI-txt.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agent Study API",
    "description": "OpenAPI documentation powered by springdoc-openapi",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:6363",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/tool/{id}/toggle": {
      "put": {
        "tags": [
          "工具管理"
        ],
        "summary": "切换工具状态",
        "description": "启用或禁用指定的工具",
        "operationId": "toggleToolStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/tool/enabled/{userId}": {
      "get": {
        "tags": [
          "工具管理"
        ],
        "summary": "获取用户启用的工具",
        "description": "从数据库查询指定用户所有启用状态的工具",
        "operationId": "getUserEnabledTools",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanListSQLTool"
                }
              }
            }
          }
        }
      }
    },
    "/tool-audit/tool/{toolName}": {
      "get": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "根据工具名称查询审计日志",
        "description": "查询指定工具的调用记录",
        "operationId": "getLogsByToolName",
        "parameters": [
          {
            "name": "toolName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanListToolAuditLog"
                }
              }
            }
          }
        }
      }
    },
    "/tool-audit/time-range": {
      "get": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "查询时间范围内的审计日志",
        "description": "查询指定时间段内的所有工具调用记录",
        "operationId": "getLogsByTimeRange",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanListToolAuditLog"
                }
              }
            }
          }
        }
      }
    },
    "/tool-audit/statistics": {
      "get": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "获取工具使用统计",
        "description": "获取指定时间范围内的工具使用统计信息",
        "operationId": "getToolUsageStatistics",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/tool-audit/failed": {
      "get": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "查询失败的工具调用",
        "description": "查询所有执行失败的工具调用记录",
        "operationId": "getFailedLogs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanListToolAuditLog"
                }
              }
            }
          }
        }
      }
    },
    "/tool-audit/count/{toolName}": {
      "get": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "统计工具调用次数",
        "description": "统计指定工具在时间范围内的调用次数",
        "operationId": "countToolUsage",
        "parameters": [
          {
            "name": "toolName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanMapStringObject"
                }
              }
            }
          }
        }
      }
    },
    "/tool-audit/conversation/{conversationId}": {
      "get": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "根据对话ID查询审计日志",
        "description": "查询指定对话中的所有工具调用记录",
        "operationId": "getLogsByConversationId",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanListToolAuditLog"
                }
              }
            }
          }
        }
      }
    },
...
    "/tool/{id}": {
      "delete": {
        "tags": [
          "工具管理"
        ],
        "summary": "删除工具",
        "description": "从数据库删除指定的工具",
        "operationId": "deleteTool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanVoid"
                }
              }
            }
          }
        }
      }
    },
    "/tool-audit/cleanup": {
      "delete": {
        "tags": [
          "工具审计管理"
        ],
        "summary": "清理历史审计日志",
        "description": "删除指定时间之前的审计日志",
        "operationId": "deleteLogsBefore",
        "parameters": [
          {
            "name": "beforeTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBeanMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultBeanVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "返回信息"
          },
          "data": {
            "type": "object",
            "description": "返回数据"
          }
        }
      },
...
      "ResultBeanMapStringObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "返回信息"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "description": "返回数据"
            },
            "description": "返回数据"
          }
        }
      },
...
      "ResultBeanListSQLTool": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "返回信息"
          },
          "data": {
            "type": "array",
            "description": "返回数据",
            "items": {
              "$ref": "#/components/schemas/SQLTool"
            }
          }
        }
      },
      "SQLTool": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "工具ID",
            "format": "int32"
          },
          "toolName": {
            "type": "string",
            "description": "工具名称"
          },
          "description": {
            "type": "string",
            "description": "工具描述"
          },
          "className": {
            "type": "string",
            "description": "工具类的完整类名(MCP对应 ServerName,USER 对应 用户id )"
          },
          "inputSchema": {
            "type": "string",
            "description": "工具参数定义(JSON格式)"
          },
          "enabled": {
            "type": "integer",
            "description": "是否启用: 1-启用, 0-禁用",
            "format": "int32"
          },
          "category": {
            "type": "string",
            "description": "工具分类"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "返回数据"
      },
      "ResultBeanListToolAuditLog": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "返回信息"
          },
          "data": {
            "type": "array",
            "description": "返回数据",
            "items": {
              "$ref": "#/components/schemas/ToolAuditLog"
            }
          }
        }
      },
      "ToolAuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "日志ID",
            "format": "int32"
          },
          "conversationId": {
            "type": "string",
            "description": "对话ID"
          },
          "requestId": {
            "type": "string",
            "description": "请求ID"
          },
          "toolName": {
            "type": "string",
            "description": "工具名称"
          },
          "methodName": {
            "type": "string",
            "description": "方法名称"
          },
          "status": {
            "type": "string",
            "description": "执行状态: STARTED, COMPLETED, FAILED"
          },
          "parameters": {
            "type": "string",
            "description": "工具参数(JSON格式)"
          },
          "result": {
            "type": "string",
            "description": "执行结果(JSON格式)"
          },
          "errorMessage": {
            "type": "string",
            "description": "错误信息"
          },
          "executionTime": {
            "type": "integer",
            "description": "执行耗时(毫秒)",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "startAt": {
            "type": "string",
            "description": "开始时间",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "description": "结束时间",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "description": "用户ID"
          },
          "ipAddress": {
            "type": "string",
            "description": "IP地址"
          }
        },
        "description": "返回数据"
      },
...
    }
  }
}